+++ src/components/spartan/SInputMaskBlock/SInputMaskBlock.vue
@@ -1,29 +1,17 @@
 <script setup lang="ts">
-import { computed } from 'vue';
 import { SInputMask, type TInputMaskProps, type TInputMaskEmits } from '../SInputMask';
 import { BlockWrapper, type TBlockWrapperProps } from '@internal';
 import type { TInputProps } from '../SInput';
+import { extractWrapperProps } from '@/helpers';
 
 defineEmits<TInputMaskEmits>();
 const props = defineProps<Partial<TBlockWrapperProps> & Partial<TInputProps> & TInputMaskProps>();
-
-const blockWrapperProps = computed(() => ({
-    label: props.label,
-    id: props.id,
-    errorText: props.errorText,
-    helpText: props.helpText,
-}));
-
-const SInputMaskProps = computed<Partial<TInputProps> & TInputMaskProps>(() => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { label, helpText, errorText, ...rest } = props;
-    return { ...rest, error: props.errorText ? Boolean(props.errorText) : props.error };
-});
+const [blockWrapperProps, inputMaskProps] = extractWrapperProps<Partial<TInputProps> & TInputMaskProps>(props);
 </script>
 
 <template>
-    <BlockWrapper wrapper="SInputMaskBlock" v-bind="blockWrapperProps">
-        <SInputMask class="w-full" v-bind="SInputMaskProps">
+    <BlockWrapper wrapper="SInputMaskBlock" v-bind="blockWrapperProps" v-slot="{ id }">
+        <SInputMask :id="id" class="w-full" v-bind="inputMaskProps">
             <template #left><slot name="left" /></template>
             <template #right><slot name="right" /></template>
         </SInputMask>